+++ src/main/java/com/decaf/domain/order/repository/OrderRepository.java
@@ -1,5 +1,6 @@
-package com.decaf.domain.order;
+package com.decaf.domain.order.repository;
 
+import com.decaf.domain.order.entity.Order;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface OrderRepository extends JpaRepository<Order, Integer> {
