+++ src/main/java/com/decaf/domain/orderItem/dto/OrderItemResponse.java
@@ -1,5 +1,7 @@
 package com.decaf.domain.orderItem.dto;
 
+import com.decaf.domain.orderItem.entity.OrderItem;
+
 import java.time.LocalDateTime;
 
 public record OrderItemResponse(
@@ -13,17 +15,17 @@ public record OrderItemResponse(
         int totalPrice,
         LocalDateTime createdAt
 ) {
-//    public OrderItemResponse(OrderItem item) {
-//        this(
-//                item.getId(),
-//                item.getOrder().getId(),
-//                item.getProduct().getId(),
-//                item.getProduct().getName(),
-//                item.getCategory(),
-//                item.getPrice(),
-//                item.getQuantity(),
-//                item.getPrice() * item.getQuantity(),
-//                item.getCreatedAt()
-//        );
-//    }
+    public OrderItemResponse(OrderItem item) {
+        this(
+                item.getId(),
+                item.getOrder().getId(),
+                item.getProduct().getId(),
+                item.getProduct().getName(),
+                item.getCategory(),
+                item.getPrice(),
+                item.getQuantity(),
+                item.getPrice() * item.getQuantity(),
+                item.getCreatedAt()
+        );
+    }
 }
